Model integrated into imms.

+++ imms/immscore/sqlite++.h
@@ -136,7 +136,7 @@ public:
     SQLQuery &operator>>(string &s);
 
     SQLQuery &load(void *data, size_t &n);
-    SQLQuery &load(void *data, unsigned n) {
+    SQLQuery &load(void *data, unsigned long long n) {
         size_t real_size = n;
         return load(data, real_size);
     };
